Move api.js
